refactor(web): split admin job management views (#26)
