+++ apps/web/src/features/auth/components/AdminRoute.tsx
@@ -2,13 +2,18 @@ import type { PropsWithChildren } from "react";
 import { Navigate, useLocation } from "react-router-dom";
 import { useSessionStore } from "@/store/session.store";
 
-export function ProtectedRoute({ children }: PropsWithChildren) {
+export function AdminRoute({ children }: PropsWithChildren) {
   const isAuthenticated = useSessionStore((state) => state.isAuthenticated);
+  const user = useSessionStore((state) => state.user);
   const location = useLocation();
   const from = `${location.pathname}${location.search}`;
 
   if (!isAuthenticated) {
     return <Navigate to="/login" replace state={{ from }} />;
+  }
+
+  if (user?.role !== "admin") {
+    return <Navigate to="/app" replace />;
   }
 
   return children;
