+++ apps/web/src/features/jobs/views/ManageJobView.tsx
@@ -1,17 +1,7 @@
-import ArrowBackRoundedIcon from "@mui/icons-material/ArrowBackRounded";
-import DeleteOutlineRoundedIcon from "@mui/icons-material/DeleteOutlineRounded";
-import SaveRoundedIcon from "@mui/icons-material/SaveRounded";
-import VisibilityOffRoundedIcon from "@mui/icons-material/VisibilityOffRounded";
-import {
-  Alert,
-  Button,
-  Checkbox,
-  FormControlLabel,
-  TextField
-} from "@mui/material";
+import { Alert } from "@mui/material";
 import { useEffect, useState } from "react";
-import type { Dispatch, FormEvent, ReactNode, SetStateAction } from "react";
-import { Link, useNavigate } from "react-router-dom";
+import type { FormEvent } from "react";
+import { useNavigate } from "react-router-dom";
 import { AppHeader } from "@/components/AppHeader";
 import { SectionCard } from "@/components/SectionCard";
 import {
@@ -25,6 +15,9 @@ import type {
   CreateJobRequestDto,
   JobDetailsResponseDto
 } from "@/api/jobs/jobs.types";
+import { JobEditorAlerts } from "@/features/jobs/components/JobEditorAlerts";
+import { JobEditorForm } from "@/features/jobs/components/JobEditorForm";
+import { JobEditorHeader } from "@/features/jobs/components/JobEditorHeader";
 import { createAsyncState, type AsyncState } from "@/lib/async/async-state";
 import { useSessionStore } from "@/store/session.store";
 
@@ -236,415 +229,30 @@ export function ManageJobView({ jobId }: ManageJobViewProps) {
       <AppHeader variant="authenticated" />
 
       <div className="mx-auto max-w-5xl px-5 py-8 sm:px-8">
-        <Link
-          to="/admin/manage-jobs"
-          className="inline-flex items-center gap-2 text-sm text-foreground-secondary transition-colors hover:text-accent-primary"
-        >
-          <ArrowBackRoundedIcon sx={{ fontSize: 18 }} />
-          Back to jobs
-        </Link>
-
-        <div className="mt-6 flex flex-col gap-4 lg:flex-row lg:items-end lg:justify-between">
-          <div>
-            <p className="font-mono text-xs tracking-[0.2em] text-foreground-tertiary">
-              BACKEND-ALIGNED JOB CONTRACT
-            </p>
-            <h1 className="mt-2 font-serif text-4xl font-semibold text-foreground">{pageTitle}</h1>
-            <p className="mt-3 max-w-3xl text-sm text-foreground-secondary">
-              This editor mirrors the backend jobs contract directly, including hide and delete
-              behavior enforced by the API.
-            </p>
-          </div>
-          {!isCreateMode ? (
-            <div className="rounded-lg border border-border bg-background-elevated px-4 py-3 text-sm text-foreground-secondary">
-              Resource ID <span className="font-medium text-foreground">{numericJobId}</span>
-            </div>
-          ) : null}
-        </div>
-
-        {!isAdmin ? (
-          <Alert severity="info" sx={{ mt: 4 }}>
-            You can inspect the job resource, but save, hide, and delete actions require an admin
-            role from the backend API.
-          </Alert>
-        ) : null}
-
-        {saveMessage ? <Alert severity="success" sx={{ mt: 4 }}>{saveMessage}</Alert> : null}
-        {errorMessage ? <Alert severity="error" sx={{ mt: 4 }}>{errorMessage}</Alert> : null}
-
-        <form onSubmit={handleSubmit} className="mt-6 space-y-6">
-          <JobFieldSection title="Core">
-            <TextField
-              label="Title"
-              value={formValues.title}
-              onChange={(event) => setStringValue(setFormValues, "title", event.target.value)}
-              fullWidth
-              required
-            />
-            <TextField
-              label="Company"
-              value={formValues.company}
-              onChange={(event) => setStringValue(setFormValues, "company", event.target.value)}
-              fullWidth
-              required
-            />
-            <TextField
-              label="Job URL"
-              value={formValues.url}
-              onChange={(event) => setStringValue(setFormValues, "url", event.target.value)}
-              fullWidth
-              required
-            />
-            <TextField
-              label="Summary"
-              value={formValues.summary}
-              onChange={(event) => setStringValue(setFormValues, "summary", event.target.value)}
-              fullWidth
-              required
-              multiline
-              minRows={3}
-            />
-            <TextField
-              label="Description"
-              value={formValues.description}
-              onChange={(event) => setStringValue(setFormValues, "description", event.target.value)}
-              fullWidth
-              required
-              multiline
-              minRows={5}
-            />
-          </JobFieldSection>
-
-          <JobFieldSection title="Source">
-            <TextField
-              label="Source name"
-              value={formValues.sourceName}
-              onChange={(event) => setStringValue(setFormValues, "sourceName", event.target.value)}
-              fullWidth
-              required
-            />
-            <TextField
-              label="Source job ID"
-              value={formValues.sourceJobId}
-              onChange={(event) => setStringValue(setFormValues, "sourceJobId", event.target.value)}
-              fullWidth
-              required
-            />
-            <TextField
-              label="Source ad reference"
-              value={formValues.sourceAdReference ?? ""}
-              onChange={(event) =>
-                setNullableStringValue(setFormValues, "sourceAdReference", event.target.value)
-              }
-              fullWidth
-            />
-            <TextField
-              label="Job refresh run ID"
-              value={formValues.jobRefreshRunId ?? ""}
-              onChange={(event) =>
-                setNullableNumberValue(setFormValues, "jobRefreshRunId", event.target.value)
-              }
-              fullWidth
-            />
-            <TextField
-              label="Raw payload JSON"
-              value={formValues.rawPayloadJson}
-              onChange={(event) =>
-                setStringValue(setFormValues, "rawPayloadJson", event.target.value)
-              }
-              fullWidth
-              required
-              multiline
-              minRows={4}
-            />
-          </JobFieldSection>
-
-          <JobFieldSection title="Location">
-            <TextField
-              label="Postcode"
-              value={formValues.postcode}
-              onChange={(event) => setStringValue(setFormValues, "postcode", event.target.value)}
-              fullWidth
-              required
-            />
-            <TextField
-              label="Location name"
-              value={formValues.locationName}
-              onChange={(event) =>
-                setStringValue(setFormValues, "locationName", event.target.value)
-              }
-              fullWidth
-              required
-            />
-            <TextField
-              label="Location display name"
-              value={formValues.locationDisplayName ?? ""}
-              onChange={(event) =>
-                setNullableStringValue(setFormValues, "locationDisplayName", event.target.value)
-              }
-              fullWidth
-            />
-            <TextField
-              label="Location area JSON"
-              value={formValues.locationAreaJson ?? ""}
-              onChange={(event) =>
-                setNullableStringValue(setFormValues, "locationAreaJson", event.target.value)
-              }
-              fullWidth
-            />
-            <TextField
-              label="Latitude"
-              value={formValues.latitude ?? ""}
-              onChange={(event) =>
-                setNullableNumberValue(setFormValues, "latitude", event.target.value)
-              }
-              fullWidth
-            />
-            <TextField
-              label="Longitude"
-              value={formValues.longitude ?? ""}
-              onChange={(event) =>
-                setNullableNumberValue(setFormValues, "longitude", event.target.value)
-              }
-              fullWidth
-            />
-          </JobFieldSection>
-
-          <JobFieldSection title="Classification and salary">
-            <TextField
-              label="Category tag"
-              value={formValues.categoryTag ?? ""}
-              onChange={(event) =>
-                setNullableStringValue(setFormValues, "categoryTag", event.target.value)
-              }
-              fullWidth
-            />
-            <TextField
-              label="Category label"
-              value={formValues.categoryLabel ?? ""}
-              onChange={(event) =>
-                setNullableStringValue(setFormValues, "categoryLabel", event.target.value)
-              }
-              fullWidth
-            />
-            <TextField
-              label="Salary min"
-              value={formValues.salaryMin ?? ""}
-              onChange={(event) =>
-                setNullableNumberValue(setFormValues, "salaryMin", event.target.value)
-              }
-              fullWidth
-            />
-            <TextField
-              label="Salary max"
-              value={formValues.salaryMax ?? ""}
-              onChange={(event) =>
-                setNullableNumberValue(setFormValues, "salaryMax", event.target.value)
-              }
-              fullWidth
-            />
-            <TextField
-              label="Salary currency"
-              value={formValues.salaryCurrency ?? ""}
-              onChange={(event) =>
-                setNullableStringValue(setFormValues, "salaryCurrency", event.target.value)
-              }
-              fullWidth
-            />
-            <TextField
-              label="Contract time"
-              value={formValues.contractTime ?? ""}
-              onChange={(event) =>
-                setNullableStringValue(setFormValues, "contractTime", event.target.value)
-              }
-              fullWidth
-            />
-            <TextField
-              label="Contract type"
-              value={formValues.contractType ?? ""}
-              onChange={(event) =>
-                setNullableStringValue(setFormValues, "contractType", event.target.value)
-              }
-              fullWidth
-            />
-          </JobFieldSection>
-
-          <JobFieldSection title="Flags and timestamps">
-            <div className="grid gap-2 md:grid-cols-3">
-              <FormControlLabel
-                control={
-                  <Checkbox
-                    checked={formValues.isFullTime}
-                    onChange={(event) =>
-                      setBooleanValue(setFormValues, "isFullTime", event.target.checked)
-                    }
-                  />
-                }
-                label="Full time"
-              />
-              <FormControlLabel
-                control={
-                  <Checkbox
-                    checked={formValues.isPartTime}
-                    onChange={(event) =>
-                      setBooleanValue(setFormValues, "isPartTime", event.target.checked)
-                    }
-                  />
-                }
-                label="Part time"
-              />
-              <FormControlLabel
-                control={
-                  <Checkbox
-                    checked={formValues.isPermanent}
-                    onChange={(event) =>
-                      setBooleanValue(setFormValues, "isPermanent", event.target.checked)
-                    }
-                  />
-                }
-                label="Permanent"
-              />
-              <FormControlLabel
-                control={
-                  <Checkbox
-                    checked={formValues.isContract}
-                    onChange={(event) =>
-                      setBooleanValue(setFormValues, "isContract", event.target.checked)
-                    }
-                  />
-                }
-                label="Contract"
-              />
-              <FormControlLabel
-                control={
-                  <Checkbox
-                    checked={formValues.isRemote}
-                    onChange={(event) =>
-                      setBooleanValue(setFormValues, "isRemote", event.target.checked)
-                    }
-                  />
-                }
-                label="Remote"
-              />
-              <FormControlLabel
-                control={
-                  <Checkbox
-                    checked={formValues.isHidden}
-                    onChange={(event) =>
-                      setBooleanValue(setFormValues, "isHidden", event.target.checked)
-                    }
-                  />
-                }
-                label="Hidden"
-              />
-              <FormControlLabel
-                control={
-                  <Checkbox
-                    checked={formValues.salaryIsPredicted ?? false}
-                    onChange={(event) =>
-                      setNullableBooleanValue(
-                        setFormValues,
-                        "salaryIsPredicted",
-                        event.target.checked
-                      )
-                    }
-                  />
-                }
-                label="Salary predicted"
-              />
-            </div>
-            <TextField
-              label="Posted at UTC"
-              value={formValues.postedAtUtc}
-              onChange={(event) =>
-                setStringValue(setFormValues, "postedAtUtc", event.target.value)
-              }
-              fullWidth
-              required
-            />
-            <TextField
-              label="Imported at UTC"
-              value={formValues.importedAtUtc}
-              onChange={(event) =>
-                setStringValue(setFormValues, "importedAtUtc", event.target.value)
-              }
-              fullWidth
-              required
-            />
-            <TextField
-              label="Last seen at UTC"
-              value={formValues.lastSeenAtUtc}
-              onChange={(event) =>
-                setStringValue(setFormValues, "lastSeenAtUtc", event.target.value)
-              }
-              fullWidth
-              required
-            />
-          </JobFieldSection>
-
-          <SectionCard className="flex flex-col gap-4 p-6 sm:flex-row sm:items-center sm:justify-between">
-            <div className="text-sm text-foreground-secondary">
-              {isCreateMode
-                ? "Create a new job resource in the backend table."
-                : "Save changes, hide the job, or delete it after it has been hidden."}
-            </div>
-            <div className="flex flex-wrap gap-3">
-              {!isCreateMode ? (
-                <Button
-                  type="button"
-                  variant="outlined"
-                  color="inherit"
-                  startIcon={<VisibilityOffRoundedIcon />}
-                  disabled={!isAdmin || isHiding || isSaving || isDeleting}
-                  onClick={() => void handleHide()}
-                >
-                  {isHiding ? "Hiding..." : "Hide job"}
-                </Button>
-              ) : null}
-              {!isCreateMode ? (
-                <Button
-                  type="button"
-                  color="error"
-                  variant="outlined"
-                  startIcon={<DeleteOutlineRoundedIcon />}
-                  disabled={!isAdmin || isDeleting || isSaving || isHiding}
-                  onClick={() => void handleDelete()}
-                >
-                  {isDeleting ? "Deleting..." : "Delete job"}
-                </Button>
-              ) : null}
-              <Button
-                type="submit"
-                variant="contained"
-                startIcon={<SaveRoundedIcon />}
-                disabled={!isAdmin || isSaving || isHiding || isDeleting}
-                sx={{
-                  bgcolor: "accent.main",
-                  "&:hover": { bgcolor: "accent.dark" }
-                }}
-              >
-                {isSaving ? "Saving..." : isCreateMode ? "Create job" : "Save changes"}
-              </Button>
-            </div>
-          </SectionCard>
-        </form>
+        <JobEditorHeader
+          isCreateMode={isCreateMode}
+          jobId={numericJobId}
+          pageTitle={pageTitle}
+        />
+        <JobEditorAlerts
+          errorMessage={errorMessage}
+          isAdmin={isAdmin}
+          saveMessage={saveMessage}
+        />
+        <JobEditorForm
+          formValues={formValues}
+          isAdmin={isAdmin}
+          isCreateMode={isCreateMode}
+          isDeleting={isDeleting}
+          isHiding={isHiding}
+          isSaving={isSaving}
+          onDelete={() => void handleDelete()}
+          onHide={() => void handleHide()}
+          onSubmit={handleSubmit}
+          setFormValues={setFormValues}
+        />
       </div>
     </div>
-  );
-}
-
-function JobFieldSection({
-  children,
-  title
-}: {
-  children: ReactNode;
-  title: string;
-}) {
-  return (
-    <SectionCard className="p-6">
-      <h2 className="font-serif text-2xl font-semibold text-foreground">{title}</h2>
-      <div className="mt-5 grid gap-4 md:grid-cols-2">{children}</div>
-    </SectionCard>
   );
 }
 
@@ -763,59 +371,4 @@ function normalizeNullableString(value: string | null): string | null {
 
   const trimmed = value.trim();
   return trimmed ? trimmed : null;
-}
-
-function setStringValue<TKey extends keyof CreateJobRequestDto>(
-  setFormValues: Dispatch<SetStateAction<CreateJobRequestDto>>,
-  key: TKey,
-  value: CreateJobRequestDto[TKey]
-) {
-  setFormValues((current) => ({
-    ...current,
-    [key]: value
-  }));
-}
-
-function setNullableStringValue<TKey extends keyof CreateJobRequestDto>(
-  setFormValues: Dispatch<SetStateAction<CreateJobRequestDto>>,
-  key: TKey,
-  value: string
-) {
-  setFormValues((current) => ({
-    ...current,
-    [key]: value === "" ? null : value
-  }));
-}
-
-function setNullableNumberValue<TKey extends keyof CreateJobRequestDto>(
-  setFormValues: Dispatch<SetStateAction<CreateJobRequestDto>>,
-  key: TKey,
-  value: string
-) {
-  setFormValues((current) => ({
-    ...current,
-    [key]: value === "" ? null : Number(value)
-  }));
-}
-
-function setBooleanValue<TKey extends keyof CreateJobRequestDto>(
-  setFormValues: Dispatch<SetStateAction<CreateJobRequestDto>>,
-  key: TKey,
-  value: boolean
-) {
-  setFormValues((current) => ({
-    ...current,
-    [key]: value
-  }));
-}
-
-function setNullableBooleanValue<TKey extends keyof CreateJobRequestDto>(
-  setFormValues: Dispatch<SetStateAction<CreateJobRequestDto>>,
-  key: TKey,
-  value: boolean
-) {
-  setFormValues((current) => ({
-    ...current,
-    [key]: value
-  }));
 }
